requirements.txt
python-telegram-bot==21.6
faster-whisper==1.0.3
yt-dlp
